1. 加入 northwind.db，改用 sqlite db 2. 換掉 nuget 套件( sqlserver -> sqlite) 3. 變更 ConnectionString 4. 重新生成 Models，同時改掉因sqlite反向工程而生成的不同型別(UnitPrice) 5. 把新建狀態功能的拿掉(只動到ProductService)

+++ NorthwindMVC/Models/Order.cs
@@ -19,7 +19,7 @@ public partial class Order
 
     public int? ShipVia { get; set; }
 
-    public decimal? Freight { get; set; }
+    public int? Freight { get; set; }
 
     public string? ShipName { get; set; }
