+++ NorthwindMVC/Models/OrderDetail.cs
@@ -9,11 +9,11 @@ public partial class OrderDetail
 
     public int ProductId { get; set; }
 
-    public decimal UnitPrice { get; set; }
+    public double UnitPrice { get; set; }
 
-    public short Quantity { get; set; }
+    public int Quantity { get; set; }
 
-    public float Discount { get; set; }
+    public double Discount { get; set; }
 
     public virtual Order Order { get; set; } = null!;
